FIX: 환전 잔액 Double 자료형으로 바꿈

+++ backend/src/main/java/com/noah/backend/domain/exchange/dto/requestDto/ExchangeReqDto.java
@@ -11,5 +11,5 @@ public class ExchangeReqDto {
     Long travelId;
     String currency;        // 통화
     int amount;          // 환전금액
-    int exchangeAmount;  // 외화금액
+    Double exchangeAmount;  // 외화금액
 }
